requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
apscheduler==3.10.4
python-dateutil==2.8.2
aiohttp==3.9.1